fix: return Account instead of generator (#28)
* fix: return Account instead of generator

After some changes to the API, AC.account() ended up being a generator,
and returned a generator object regardless of the value of stream.

This commit replaces 'yield from' with 'return', ensuring that
AC.account() remains a regular function and returns an Account object
when stream=False, and a generator object when stream=True.

* docs: update CHANGELOG